# packages/contracts/openapi.yaml
openapi: 3.1.1
info:
  title: SAMTEC Security System API
  version: "0.1.0"
  summary: Biometric attendance and payroll for security companies.
  description: |
    This document is the **API contract**: the single agreement between the
    frontend (`apps/web`) and the backend (`apps/api`).

    **Golden rule:** change this file first, in its own pull request, and get
    both developers to approve it. Only then write the code. See
    `docs/guides/05-api-contract-workflow.md`.

    ## Conventions used by every endpoint

    - **Money** is an integer number of **pesewas** (GHS 1.00 = 100 pesewas).
      Money field names end in `Pesewas`.
    - **Timestamps** are UTC in ISO 8601 format, for example
      `2026-09-15T06:00:00Z`. The dashboard shows them in Ghana time
      (Africa/Accra).
    - **Calendar dates** without a time use `YYYY-MM-DD`.
    - **IDs** are UUIDs (version 7).
    - **Errors** use Problem Details (RFC 9457) with the content type
      `application/problem+json`.
    - **Lists** use cursor pagination. Send the `nextCursor` from one page as
      the `cursor` of the next request. `nextCursor: null` means there are no
      more pages.
  contact:
    name: Alpha Team
    url: https://github.com/ALPHA-TEAM-dev/samtec-security-system
servers:
  - url: /api/v1
    description: >-
      The API's own address, for example http://localhost:3000/api/v1 during
      development. The dashboard sets the full address in its configuration.
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Sign in with email and password
      description: |
        On success, the response body holds a short-lived **access token**
        (15 minutes) and the API sets a long-lived **refresh token** in an
        `HttpOnly` cookie named `samtec_refresh`. JavaScript cannot read that
        cookie, which protects it from cross-site scripting.

        Accounts with the `ADMIN` or `HR_PAYROLL` role must use two-factor
        authentication. The response `status` tells the dashboard what to do
        next:

        - `AUTHENTICATED`: signed in, with no second step.
        - `TWO_FACTOR_REQUIRED`: call `POST /auth/2fa/verify` with the
          `challengeToken` and a code from the user's authenticator app.
        - `TWO_FACTOR_SETUP_REQUIRED`: the account has not set up two-factor
          authentication yet. Call `POST /auth/2fa/setup`, show the QR code,
          then call `POST /auth/2fa/enable` with the first code.

        **Protection against password guessing:** repeated failures for the
        same email are slowed down, then answered with `429` for 15 minutes.
        This happens whether or not an account exists, so the answers never
        reveal which emails have accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              supervisor:
                summary: A supervisor signs in
                value:
                  email: ama.boateng@samtec.example
                  password: correct-horse-battery-staple
      responses:
        "200":
          description: Signed in, or a second factor is required.
          headers:
            Set-Cookie:
              description: The refresh token cookie. Only sent when `status` is `AUTHENTICATED`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                authenticated:
                  summary: Signed in with no second factor needed
                  value:
                    status: AUTHENTICATED
                    accessToken: example-access-token
                    expiresInSeconds: 900
                    user:
                      id: 01927c3e-2222-7ccc-9ddd-0e0e0e0e0e02
                      email: ama.boateng@samtec.example
                      fullName: Ama Boateng
                      role: SUPERVISOR
                      twoFactorEnabled: false
                      employeeId: 01927c3e-5a4b-7c8d-9e0f-1a2b3c4d5e70
                twoFactorRequired:
                  summary: Admin or HR account needs a second factor
                  value:
                    status: TWO_FACTOR_REQUIRED
                    challengeToken: example-challenge-token
                    expiresInSeconds: 300
                twoFactorSetupRequired:
                  summary: Admin or HR account must set up two-factor authentication first
                  value:
                    status: TWO_FACTOR_SETUP_REQUIRED
                    setupToken: example-setup-token
                    expiresInSeconds: 600
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/2fa/verify:
    post:
      tags: [Auth]
      operationId: verifyTwoFactor
      summary: Finish signing in with a two-factor code
      description: |
        Send the `challengeToken` from `POST /auth/login` and the 6-digit code
        from the user's authenticator app. On success the API responds like a
        normal sign-in and sets the refresh token cookie.

        Rules the API enforces:

        - A challenge token expires after 5 minutes, works only once, and
          belongs to the one account that signed in.
        - After 5 wrong codes the challenge is cancelled, and the user must
          sign in with their password again.
        - A code that was already accepted cannot be used a second time, even
          within its 30-second window (replay protection).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTwoFactorRequest"
      responses:
        "200":
          description: Signed in.
          headers:
            Set-Cookie:
              description: The refresh token cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/2fa/setup:
    post:
      tags: [Auth]
      operationId: startTwoFactorSetup
      summary: Start setting up two-factor authentication
      description: |
        Send the `setupToken` from `POST /auth/login`. The API creates a new
        secret for the account and returns it as an `otpauth://` link, which
        the dashboard shows as a QR code for the authenticator app. Calling
        this again replaces the secret, so only the newest QR code works.

        The setup token expires after 10 minutes and belongs to one account.
        Two-factor authentication stays off until `POST /auth/2fa/enable`
        succeeds.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorSetupRequest"
      responses:
        "200":
          description: The new secret, ready to show as a QR code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorSetup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/2fa/enable:
    post:
      tags: [Auth]
      operationId: enableTwoFactor
      summary: Finish setting up two-factor authentication
      description: >-
        Send the `setupToken` and the first 6-digit code from the authenticator
        app. A correct code proves the app was set up properly. The API then
        switches two-factor authentication on, uses up the setup token and
        signs the user in. The rules of `POST /auth/2fa/verify` apply here too:
        5 wrong codes cancel the setup, and a code works only once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableTwoFactorRequest"
      responses:
        "200":
          description: Two-factor authentication is on, and the user is signed in.
          headers:
            Set-Cookie:
              description: The refresh token cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Get a new access token using the refresh cookie
      description: |
        The browser sends the `samtec_refresh` cookie automatically. The API
        replaces it with a new one on every call (rotation).

        - **Reuse detection:** a refresh token that was already replaced must
          never arrive again. If it does, someone may have copied it, so the
          API signs that user out everywhere by revoking all of their refresh
          tokens.
        - **Origin check:** the request's `Origin` header must be one of the
          dashboard addresses in the API's `CORS_ORIGINS` setting. Otherwise
          the API answers `403`.
      security:
        - refreshCookie: []
      responses:
        "200":
          description: A new access token.
          headers:
            Set-Cookie:
              description: The rotated refresh token cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Sign out and revoke the refresh token
      description: >-
        Like `POST /auth/refresh`, the request's `Origin` header must be one of
        the dashboard addresses in `CORS_ORIGINS`, so another website cannot
        sign the user out.
      security:
        - refreshCookie: []
      responses:
        "204":
          description: Signed out. The refresh cookie is cleared, even if it was already invalid.
          headers:
            Set-Cookie:
              description: Clears the refresh token cookie.
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the signed-in user
      responses:
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The access token from `POST /auth/login`, sent in the header
        `Authorization: Bearer <token>`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: samtec_refresh
      description: >-
        The refresh token cookie set by the API, with the attributes
        `HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth; Max-Age=604800`.
        JavaScript cannot read it, it lasts 7 days, and the browser sends it
        only over HTTPS, only to the auth endpoints, and never with requests
        started by another website. `SameSite=Strict` needs the dashboard and
        the API on the same site, as described in
        docs/plan/03-system-architecture.md.
...
  responses:
    BadRequest:
      description: The request is invalid. The `errors` list says which fields are wrong.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            validationError:
              summary: A field failed validation
              value:
                type: "urn:samtec:problem:validation-error"
                title: Validation failed
                status: 400
                detail: One or more fields are invalid.
                instance: /api/v1/employees
                traceId: 01927c3e-9f00-7abc-8def-000000000001
                errors:
                  - path: ghanaCardNumber
                    message: Must look like GHA-123456789-0.
    Unauthorized:
      description: No valid access token was sent, or it has expired. Refresh the session or sign in again.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            notSignedIn:
              summary: Missing or expired token
              value:
                type: about:blank
                title: Unauthorized
                status: 401
                detail: Sign in to continue.
                traceId: 01927c3e-9f00-7abc-8def-000000000002
    InvalidCredentials:
      description: >-
        The sign-in details are wrong. The message never says whether the email
        or the password was wrong, so attackers cannot discover which emails
        have accounts.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            wrongPassword:
              summary: Wrong email or password
              value:
                type: about:blank
                title: Unauthorized
                status: 401
                detail: Email or password is incorrect.
                traceId: 01927c3e-9f00-7abc-8def-000000000003
    Forbidden:
      description: You are signed in, but your role is not allowed to do this.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            wrongRole:
              summary: Role not allowed
              value:
                type: about:blank
                title: Forbidden
                status: 403
                detail: Your role does not allow this action.
                traceId: 01927c3e-9f00-7abc-8def-000000000004
...
    TooManyRequests:
      description: Too many attempts in a short time. Wait before trying again.
      headers:
        Retry-After:
          description: How many seconds to wait before trying again.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            slowDown:
              summary: Rate limit reached
              value:
                type: about:blank
                title: Too Many Requests
                status: 429
                detail: Too many attempts. Try again in 60 seconds.
                traceId: 01927c3e-9f00-7abc-8def-000000000007
  schemas:
...
    ProblemDetails:
      type: object
      description: >-
        A standard error response (RFC 9457). Show `detail` to the user, and
        include `traceId` when reporting a bug so the backend developer can find
        the exact request in the logs.
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri-reference
          description: >-
            The kind of error. `about:blank` means a plain HTTP error.
            Values starting with `urn:samtec:problem:` are SAMTEC-specific.
          examples: ["urn:samtec:problem:validation-error"]
        title:
          type: string
          description: A short summary of the error type.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation of this specific error.
        instance:
          type: string
          format: uri-reference
          description: The request path that failed.
        traceId:
          type: string
          description: The ID of the request, also written to the server logs.
        errors:
          type: array
          description: Present on validation errors, with one entry per invalid field.
          items:
            $ref: "#/components/schemas/ValidationIssue"
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
          description: The invalid field as a dot path, for example `phone` or `address.city`.
        message:
          type: string
          description: What is wrong, written for the user.
    UserRole:
      type: string
      description: |
        What a user account is allowed to do.
        - `ADMIN` — full access, including user management and approving payroll.
        - `HR_PAYROLL` — manages employees and prepares payroll runs, but cannot approve a run they prepared.
        - `SUPERVISOR` — manages attendance and rosters for their assigned sites.
        - `GUARD` — sees only their own attendance and payslips.
      enum: [ADMIN, HR_PAYROLL, SUPERVISOR, GUARD]
    CurrentUser:
      type: object
      required: [id, email, fullName, role, twoFactorEnabled, employeeId]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        twoFactorEnabled:
          type: boolean
        employeeId:
          type: [string, "null"]
          format: uuid
          description: The linked employee record for guards and supervisors, or `null` for office-only accounts.
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
          writeOnly: true
    LoginResponse:
      description: >-
        A finished sign-in, a request for a two-factor code, or a request to
        set up two-factor authentication. Check `status` to tell them apart.
      oneOf:
        - $ref: "#/components/schemas/AuthenticatedSession"
        - $ref: "#/components/schemas/TwoFactorChallenge"
        - $ref: "#/components/schemas/TwoFactorSetupRequired"
      discriminator:
        propertyName: status
        mapping:
          AUTHENTICATED: "#/components/schemas/AuthenticatedSession"
          TWO_FACTOR_REQUIRED: "#/components/schemas/TwoFactorChallenge"
          TWO_FACTOR_SETUP_REQUIRED: "#/components/schemas/TwoFactorSetupRequired"
    AuthenticatedSession:
      type: object
      required: [status, accessToken, expiresInSeconds, user]
      properties:
        status:
          type: string
          const: AUTHENTICATED
        accessToken:
          type: string
          description: Send it in the `Authorization` header. Keep it in memory only, never in localStorage.
        expiresInSeconds:
          type: integer
          minimum: 1
          examples: [900]
        user:
          $ref: "#/components/schemas/CurrentUser"
    TwoFactorChallenge:
      type: object
      required: [status, challengeToken, expiresInSeconds]
      properties:
        status:
          type: string
          const: TWO_FACTOR_REQUIRED
        challengeToken:
          $ref: "#/components/schemas/OneTimeToken"
        expiresInSeconds:
          type: integer
          minimum: 1
          examples: [300]
    TwoFactorSetupRequired:
      type: object
      required: [status, setupToken, expiresInSeconds]
      properties:
        status:
          type: string
          const: TWO_FACTOR_SETUP_REQUIRED
        setupToken:
          $ref: "#/components/schemas/OneTimeToken"
        expiresInSeconds:
          type: integer
          minimum: 1
          examples: [600]
    OneTimeToken:
      type: string
      minLength: 1
      maxLength: 512
      description: >-
        Proves the password was correct. It expires quickly, works for one
        account only, and is used up by the step it was issued for. Keep it in
        memory only.
    TwoFactorCode:
      type: string
      pattern: '^\d{6}$'
      description: The 6-digit code from the authenticator app.
      examples: ["492039"]
    VerifyTwoFactorRequest:
      type: object
      additionalProperties: false
      required: [challengeToken, code]
      properties:
        challengeToken:
          $ref: "#/components/schemas/OneTimeToken"
        code:
          $ref: "#/components/schemas/TwoFactorCode"
    TwoFactorSetupRequest:
      type: object
      additionalProperties: false
      required: [setupToken]
      properties:
        setupToken:
          $ref: "#/components/schemas/OneTimeToken"
    TwoFactorSetup:
      type: object
      description: >-
        The new two-factor secret. Both fields contain the secret itself, so
        show them only on the setup screen, and never store or log them.
      required: [otpauthUri, manualEntryKey]
      properties:
        otpauthUri:
          type: string
          description: Show this link as a QR code for the authenticator app to scan.
          examples: ["otpauth://totp/SAMTEC:ama.boateng%40samtec.example?secret=EXAMPLESECRET&issuer=SAMTEC"]
        manualEntryKey:
          type: string
          description: The same secret as text, for people who cannot scan the QR code.
          examples: [EXAMPLESECRET]
    EnableTwoFactorRequest:
      type: object
      additionalProperties: false
      required: [setupToken, code]
      properties:
        setupToken:
          $ref: "#/components/schemas/OneTimeToken"
        code:
          $ref: "#/components/schemas/TwoFactorCode"
    AccessTokenResponse:
      type: object
      required: [accessToken, expiresInSeconds]
      properties:
        accessToken:
          type: string
        expiresInSeconds:
          type: integer
          minimum: 1
          examples: [900]
...